// src/app/models/CopiarSucursal.ts
export class CopiarSucursal
{
    Nombre: string|undefined;
    Distrito: string|undefined;
    Canton: string|undefined;
    Provincia: string|undefined;
    FechaApertura: string|undefined;
    Horario: string|undefined;
    Capacidad: number|undefined;
    CedulaAdministrador: string|undefined;
    SucursalPorCopiar: string|undefined;
}